+++ espforge/src/parse/processor.rs
@@ -1,10 +1,11 @@
 use anyhow::Result;
 use serde_yaml_ng::Value;
-use ProjectModel;
+
+use crate::parse::model::ProjectModel;
 
 pub trait SectionProcessor {
     fn section_key(&self) -> &'static str;
-    fn priority(&self) -> u8 { 50 } // Default priority
+    fn priority(&self) -> u32 { 50 } // Default priority
     fn process(&self, content: &Value, model: &mut ProjectModel) -> Result<()>;
 }
